+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 # Upcoming
 - fix: Counterexamples - fix an integer parsing bug and correctly extract datatype and field names 
+- feat: New option `-diagnosticsFormat:json` to print Dafny error messages as JSON objects (one per line). (https://github.com/dafny-lang/dafny/pull/2363)
 
 # 3.7.3